webapp/requirements.txt
torch==1.13.1
torchvision==0.14.1

# Core AI + Vision
transformers==4.49.0
openai==1.91.0
opencv-python-headless==4.11.0.86
pillow==10.4.0
RapidFuzz==3.13.0

# Google AI & Cloud
google-genai==1.8.0
google-cloud-storage==2.18.2
google-cloud-firestore==2.19.0

# HTTP / Auth / Utils
requests==2.32.3
httpx==0.28.1
python-dotenv==1.0.1
PyJWT==2.10.1
certifi==2025.6.15
sniffio==1.3.1
six==1.16.0
idna==3.7
pyyaml==6.0.2

# Real-ESRGAN & dependencies
realesrgan==0.3.0
basicsr==1.4.2
gfpgan==1.3.8

# PaddleOCR
paddleocr==3.0.2
paddlepaddle==3.0.0

# OpenSearch
opensearch-py==2.8.0
cloudscraper==1.2.71
beautifulsoup4==4.13.4
numpy==1.26.4